split running flag utils
