+++ src/utils/running.rs
@@ -1,0 +1,30 @@
+use anyhow::{anyhow, Result};
+use std::sync::atomic::AtomicBool;
+use std::sync::{atomic, Arc, OnceLock};
+
+/// The running flag, used to check if the user aborted.
+pub fn running_flag() -> &'static Arc<AtomicBool> {
+    static RUNNING: OnceLock<Arc<AtomicBool>> = OnceLock::new();
+    RUNNING.get_or_init(|| Arc::new(AtomicBool::new(true)))
+}
+
+/// Check whether the program should continue running.
+pub fn is_running() -> bool {
+    running_flag().load(atomic::Ordering::Relaxed)
+}
+
+/// Check whether the user asked to abort. It's the same as `!running()`, but return a Result.
+pub fn user_aborted() -> Result<()> {
+    match is_running() {
+        true => Ok(()),
+        false => Err(anyhow!("aborted")),
+    }
+}
+
+/// Return a static string, suitable for displaying, regarding the state of some computation
+/// that might have been aborted.
+pub fn aborted(cond: bool) -> &'static str {
+    (cond && !is_running())
+        .then_some(" (partial, aborted)")
+        .unwrap_or_default()
+}
